+++ ext/Shared/WeaponLists/CustomWeaponsAssault.lua
@@ -1,0 +1,35 @@
+CustomWeaponsAssault = {
+	-- Assault Rifles
+	'AEK971',
+	'AN94',
+	'AK74M',
+	-- 'SteyrAug',
+	-- 'F2000',
+	-- 'FAMAS',
+	-- 'G3A3',
+	-- 'KH2002',
+	-- 'L85A2',
+	'M16A4',
+	'M416',
+	-- 'SCAR-L',
+
+	-- Shotguns
+	-- 'DAO-12',
+	-- 'Jackhammer',
+	-- 'Saiga20',
+	-- 'SPAS12',
+	-- 'USAS-12',
+	-- 'M1014',
+	-- '870M',
+
+	-- PDW
+	-- 'ASVal',
+	-- 'MP7',
+	-- 'P90',
+	-- 'PP-19',
+	-- 'PP2000',
+	-- 'UMP45',
+	-- 'MP5K',
+	-- 'MagpulPDR',
+	-- 'Crossbow'
+}