+++ ext/Shared/WeaponLists/CustomWeaponsPistols.lua
@@ -1,0 +1,19 @@
+CustomWeaponsPistols = {
+	-- Pisotls
+	-- 'Glock17',
+	-- 'Glock17_Silenced',
+	'Glock18',
+	-- 'Glock18_Silenced',
+	'M1911_Lit',
+	-- 'M1911_Silenced',
+	'M1911_Tactical',
+	-- 'M9',
+	'M9_Silenced',
+	-- 'M9_Tactical',
+	-- 'M93R',
+	'MP412Rex',
+	'Taurus44',
+	-- 'MP443',
+	-- 'MP443_Tactical',
+	-- 'MP443_Silenced',
+}